Crate janetrs_macros

Macros

check_janet_version

Checks the specified version of Janet is equal of the used Janet. Emits a boolean.

declare_janet_mod

Declare a Janet module.

Attribute Macros

cjvg

Conditional Janet Version Gate

janet_fn

Macro that creates from a high-level Janet function (fn(&mut [Janet]) -> Janet) to the kind of function the Janet C API is expecting (fn(i32, *mut janetrs::lowlevel::Janet) -> janetrs::lowlevel::Janet) with a modified name with _c postfix.

janet_version

Conditional Janet Version Gate