janetrs_macros

Attribute Macro janet_version

#[janet_version]

Conditional Janet Version Gate

Usage: #[janet_version(<MIN_VERSION>, [MAX_VERSION])] where MIN_VERSION and MAX_VERSION are string literals.

A macro da conditionally includes the input if the version of Janet is bigger or equal to the passed minimal version and smaller than the passed maximum version.

That means that the range is open in the maximum version: [MIN_VERSION, MAX_VERSION).